requirements.txt
fastapi
langchain
langchain-openai
uvicorn
firecrawl
pdfkit
google-cloud-documentai
langchain-community
pypdf
docx2txt
openpyxl
SQLAlchemy
psycopg[binary]
pgvector

python-dotenv